# feat: Use lazy initialization for msg, chat, and from