Reduce N number of ternary comparison down to a single boolean comparison, if the property has been accessed before.