@KnorpelSenf 2025-08-16
I don't find it obvious that this would make things faster. Sure, we have a few property access operations and bullish checks. On the flipside:
- we increase GC pressure (more references)
- we increase memory consumption
- we optimise for an uncommon case: most bots perform a low, constant number of reads on the context object and they usually cache the values on the stack anyway if they access them several times

I'd at least like to better understand these changes. Perhaps we should bench them for 1M context object creations with `{0..3}` property access calls each to find out how this changes things?