@KnightNiwrem 2025-08-16
Some context: This is an alternative way to address, not so much an performance-focused PQ

> Also, do you think we should deduplicate the ctx.msg and ctx.msgId getters so that the ctx.msg getter really isn't computed repeatedly?

The most straightforward way was to simply change all `this.msgId` with `msg?.message_id`, given that we already do `const msg = this.msg`.

However, this creates some degree of cognitive load to remember creating some objects when we intend to access the object more than once in the ctx methods.

The same also applies to plugin writers and developers, if they are interested in such optimisations. It's also not as pretty - we do want to simply do `ctx.msgId` if we can. Especially broadcast bots doing `ctx.copyMessage` or `ctx.forwardMessage`.

This alternative appears accomplish the same thing as the most straightforward way as well. After `this.msgId` is accessed, any further calls to `this.msg?.is_topic_message`, for example, no longer performs duplicate computation.

It covers all methods that don't currently do `const msg = this.msg` for any future updates where it tries to access more properties of msg.

And it incidentally covers plugin writers and developers so they simply need not think about such optimisation, and may write however is most intuitive.

That said, a perf comparison would still be interesting. Just wanted to clarify the thoughts behind this approach.