@KnightNiwrem 2025-08-16
Re: GC and memory

Not really expecting much impact in those places. A Context must contain an Update. It cannot not have an Update. So either the Context has 2 references to Message (and other) or it has 0 references to Message (because it is destroyed).

There is no case where Context has a reference to Message via the new field, yet the Message is not referenced by Context indirectly via Update.

I'm expecting the reference to be a fixed byte size value, since it is simply a shallow reference.

The biggest concern for memory probably would be when Context is JSONified and stored. Then the fields would need to be dereferenced and deeply copied. We can probably workaround that for our plugins, but it would still be a concern for JSON.stringify done externally.