@KnightNiwrem 2025-08-16
Worsens performance based on benchmarks. Guess we have to just do it the original way

Baseline from v1.38.0
```
| benchmark                               | time/iter (avg) |        iter/s |      (min … max)      |      p75 |      p99 |     p995 |
| --------------------------------------- | --------------- | ------------- | --------------------- | -------- | -------- | -------- |
| Bench calls to ctx.copyMessage never    |          2.6 µs |       391,400 | (  1.6 µs …  12.8 ms) |   1.9 µs |   4.7 µs |   6.5 µs |
| Bench calls to ctx.copyMessage once     |          1.8 µs |       554,400 | (  1.8 µs …   1.9 µs) |   1.8 µs |   1.9 µs |   1.9 µs |
| Bench calls to ctx.copyMessage twice    |          1.8 µs |       551,400 | (  1.8 µs …   2.0 µs) |   1.8 µs |   2.0 µs |   2.0 µs |
| Bench calls to ctx.copyMessage thrice   |          1.8 µs |       559,300 | (  1.7 µs …   1.9 µs) |   1.8 µs |   1.9 µs |   1.9 µs |
```

With lazy init
```
| benchmark                               | time/iter (avg) |        iter/s |      (min … max)      |      p75 |      p99 |     p995 |
| --------------------------------------- | --------------- | ------------- | --------------------- | -------- | -------- | -------- |
| Bench calls to ctx.copyMessage never    |          8.6 µs |       116,100 | (  5.1 µs …  13.0 ms) |   6.4 µs |  13.9 µs |  29.1 µs |
| Bench calls to ctx.copyMessage once     |          5.5 µs |       180,400 | (  5.1 µs …   6.5 µs) |   6.2 µs |   6.5 µs |   6.5 µs |
| Bench calls to ctx.copyMessage twice    |          5.3 µs |       189,100 | (  5.1 µs …   6.3 µs) |   5.2 µs |   6.3 µs |   6.3 µs |
| Bench calls to ctx.copyMessage thrice   |          5.3 µs |       190,500 | (  5.1 µs …   5.4 µs) |   5.3 µs |   5.4 µs |   5.4 µs |
```