@KnightNiwrem 2025-08-16
Benchmarking source:

```ts
const BOT_TOKEN = '';
import { Bot, type Context } from 'change this between local build and remote release';

function createUpdate(): Context['update'] {
  return {
    "update_id": 908295308,
    "message": {
      "message_id": 9396,
      "from": {
        "id": 41284431,
        "is_bot": false,
        "first_name": "KnightNiwrem",
        "username": "Knightniwrem",
        "language_code": "en"
      },
      "chat": {
        "id": 41284431,
        "first_name": "KnightNiwrem",
        "username": "Knightniwrem",
        "type": "private"
      },
      "date": 1755357026,
      "text": "hi"
    }
  };
}

const benchMiddlewares = {
  'never': () => {},
  'once': async (ctx: Context) => { await ctx.copyMessage('1') },
  'twice': async (ctx: Context) => { await ctx.copyMessage('2'); await ctx.copyMessage('2'); },
  'thrice': async (ctx: Context) => { await ctx.copyMessage('3'); await ctx.copyMessage('3'); await ctx.copyMessage('3'); },
}
let currentBenchFn: Function;

const bot = new Bot(BOT_TOKEN, {
  botInfo: {
    "id": 1561644845,
    "is_bot": true,
    "first_name": "Heroku Experiment",
    "username": "he_exp_bot",
    "can_join_groups": true,
    "can_read_all_group_messages": false,
    "supports_inline_queries": true,
    "can_connect_to_business": false,
    "has_main_web_app": true
  }
});

// @ts-expect-error
bot.api.config.use(() => {
  return { ok: true, result: true };
});

bot.on('message:text', async ctx => {
  await currentBenchFn(ctx);
});

for (const [benchKey, benchFn] of Object.entries(benchMiddlewares)) {
  currentBenchFn = benchFn;
  Deno.bench(`Bench calls to ctx.copyMessage ${benchKey}`, async () => {
    await bot.handleUpdate(createUpdate());
  });
}
```